+++ ml/context/aliases.go
@@ -17,6 +17,7 @@
 package context
 
 // Aliases to GoMLX basic types.
+//
 // Usually one would simply import the full `graph` with a period (".") so
 // the ops don't need a package qualifier. But we can't do this in `context`
 // package because we have conflicting symbols (`Exec`, `NewExec`).